removing the reserved word Resource from the library and change the library name on composer.json
