+++ lib/OpenCloud/CloudMonitoring/Asset/MetricDataPoint.php
@@ -15,22 +15,8 @@
  * limitations under the License.
  */
 
-namespace OpenCloud\CloudMonitoring\Resource;
+namespace OpenCloud\CloudMonitoring\Asset;
 
-/**
- * Changelog class.
- */
-class Changelog extends ReadOnlyResource
+class MetricDataPoint extends Metric
 {
-    private $id;
-    private $timestamp;
-    private $entity_id;
-    private $alarm_id;
-    private $check_id;
-    private $state;
-    private $analyzed_by_monitoring_zone_id;
-
-    protected static $json_name = 'changelogs/alarms';
-    protected static $json_collection_name = 'values';
-    protected static $url_resource = 'changelogs/alarms';
 }
