+++ lib/OpenCloud/CloudMonitoring/Asset/NotificationPlan.php
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-namespace OpenCloud\CloudMonitoring\Resource;
+namespace OpenCloud\CloudMonitoring\Asset;
 
 class NotificationPlan extends AbstractResource
 {
